chore: rename components folder to temp
