+++ src/app/@yourLibrary/_components/Header.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { Tooltip } from '@/components/tooltip'
+import { Tooltip } from '@/components/temp-tooltip'
 import { useYourLibraryStore } from '@/lib/stores/your-library.store'
 import { FiSearch } from 'react-icons/fi'
 import { GoArrowLeft, GoArrowRight } from 'react-icons/go'
